Refactor e2e cases.
