+++ test/e2e/e2e_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package test
+package e2e
 
 import (
 	"testing"
