+++ test/e2e/predicates.go
@@ -1,0 +1,79 @@
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package e2e
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"k8s.io/api/core/v1"
+	"k8s.io/kubernetes/pkg/scheduler/algorithm"
+)
+
+var _ = Describe("Predicates E2E Test", func() {
+	It("NodeAffinity", func() {
+		context := initTestContext()
+		defer cleanupTestContext(context)
+
+		slot := oneCPU
+		nodeName, rep := computeNode(context, oneCPU)
+		Expect(rep).NotTo(Equal(0))
+
+		affinity := &v1.Affinity{
+			NodeAffinity: &v1.NodeAffinity{
+				RequiredDuringSchedulingIgnoredDuringExecution: &v1.NodeSelector{
+					NodeSelectorTerms: []v1.NodeSelectorTerm{
+						{
+							MatchFields: []v1.NodeSelectorRequirement{
+								{
+									Key:      algorithm.NodeFieldSelectorKeyNodeName,
+									Operator: v1.NodeSelectorOpIn,
+									Values:   []string{nodeName},
+								},
+							},
+						},
+					},
+				},
+			},
+		}
+
+		job := createJob(context, "na-job", 1, 1, "nginx", slot, affinity)
+		err := waitJobReady(context, job.Name)
+		Expect(err).NotTo(HaveOccurred())
+
+		pods := getPodOfJob(context, "na-job")
+		for _, pod := range pods {
+			Expect(pod.Spec.NodeName).To(Equal(nodeName))
+		}
+	})
+
+	It("Hostport", func() {
+		context := initTestContext()
+		defer cleanupTestContext(context)
+
+		nn := clusterNodeNumber(context)
+
+		containers := createContainers("nginx", oneCPU, 28080)
+		job := createJobWithOptions(context, "kube-batch", "qj-1", int32(nn), int32(nn*2), nil, containers)
+
+		err := waitTasksReady(context, job.Name, nn)
+		Expect(err).NotTo(HaveOccurred())
+
+		err = waitTasksNotReady(context, job.Name, nn)
+		Expect(err).NotTo(HaveOccurred())
+	})
+})
